refactor: add filtration module - move imports
